backend/requirements.txt
# Smart Commercial Calculator - Backend Dependencies

# Web Framework (開發/測試用)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AI Agent
anthropic>=0.43.0

# Data Validation
pydantic>=2.5.0

# Environment Variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Code Quality
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Packaging
pyinstaller>=6.3.0